server/pyproject.toml
[tool.poetry]
name = "finodays-server"
version = "0.1.0"
description = ""
authors = ["Ilin Ivan <buran05@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.22"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
psycopg2-binary = "^2.9.9"
werkzeug = "^3.0.0"
httpx = "^0.25.0"
pytest = "^7.4.3"
pandas = "2.0.3"
gunicorn = "^21.2.0"
requests = "^2.31.0"
tinkoff-investments = "^0.2.0b62"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
